ventura: refresh DTS patch and use
Summary:
Refresh the Ventura DTS from 'v5' in Lore.  Add a mechanism
to override upstream machine.conf files and leverage this to set the
kernel DTS.

Usually the DTS is set in the machine.conf.  The upstream version
for ventura needs to specify the "aspeed-evb" right now because the
ventura DTS is unmerged.  Therefore we need to override this
downstream when we apply the patch.

Test Plan:
`bitbake linux-aspeed -ne` and confirm KERNEL_DEVICETREE
is set appropriately.

Reviewed By: peterdelevoryas

Differential Revision: D68648791

fbshipit-source-id: fbe62046bf9a2969e3236a518182bb71b95e2bfb