openbmc-gpio: Update to work with yocto <5 and 5+
Summary:
As a preparation step to upgrade yocto to lf-master on wedge100, and the new `UNPACKDIR` requirement,
this diff moves all files supplied by the `openbmc-gpio` package (including all per-platform .bbappends)
into the root `files` directory. it also updates `UNPACKDIR` as required by yocto 5+.

This is needed since yocto <5 and 5+ put the files in different directories even given the same definitions of `S`, `WORKDIR`, `UNPACKDIR` (new to yocto 5+) and `B`. Having the files in the root dir makes it possible for the same recipe to work on both yocto <5 and 5+.

This diff should be a no-op as far as the final image is concerned.

With this dif, `openbmc-gpio` is buildable on both `lf-dunfell` and `-lf-master`.

Notes:
1. The linter finds "problems" in some python files, but this diff just moves them - if it worked before, it'll work now.
2. On lighning, the same files were both in `files` and `files/openbmc-gpio-1`, thus the many deleted files on this platform, which are actually moved from `files/openbmc-gpio-1` to `files` (one of them with tiny differences).

Test Plan:
Build wedge100, see that the correct (the one from the
bbappend) file is installed as `soc_gpio_table.py`.
Watch CI.

Reviewed By: mikechoifb

Differential Revision:
D68535088

Privacy Context Container: L1188897

fbshipit-source-id: 3edd952b37bcb05cd7034f9f83cebb1d99e995ff