wedge100: Upgrade yocto to lf-master
Summary:
Also remove a patch to `agetty` that is included in the
upstream version of `util-linux` we're upgrading to. This can't be a
separate diff since the build fails without it (`util-linux` version
2.30 is not in yocto anymore), and we don't want to remove the patch
while still on the older yocto.

Test Plan: CI, tested on a lab device, all tests from `tests2` pass.

Reviewed By: mikechoifb

Differential Revision:
D68350008

Privacy Context Container: L1188897

fbshipit-source-id: 536097bfb770c68c23b96b19fef7af0aa2cff2dd