wedge100: Remove failing kvstore test
Summary:
The test checks for the existence of 2 keys, but on many
wedge100 systems they are not present.

There's no code that creates `fru1_restart_cause` at all (which causes
conveyor to fail on nodes that don't have it).

For `internal_switch_config`, it is created if some REST API endpoints
are hit, and thus unusable for a test.

This diff removes both keys.

Test Plan: CI

Reviewed By: doranand

Differential Revision: D68745896

fbshipit-source-id: 782ee5d470cf7e1e924e885b82ae24536dea42cd