brycecanyon: Remove support for LED/health error on SEL
Summary:
This is not really done by any other platform so remove
it to stay consistent. Also, considering its based on
maintaining a huge list of "allow-list" SELs and setting
event on any SEL we might end up with a lot of false positives.

Remove it to make the code a bit more readable and
easier to maintain. Tooling will not be affected since
the SELs will still be available via log-util.

Test Plan: Bryce canyon builds.

Reviewed By: waffle2k

Differential Revision: D68449476

fbshipit-source-id: 53b37369d6630343dd09bf57df0987d1868e43d8