Run fans in manual mode on boot if config file exists
Summary:
As part of smoke testing, we want to always run the fans at 80 percent.

This is gated by the existence of a config file in the `/home/root/` path.

Upon startup, the timer will run the script which then sets the fans to the specified values, or 80% if not stated.

If the config file doesnt exist, then the fans will be run in auto mode

Test Plan:
This doesnt work in qemu, so testing on a real bmc

Check the fans See that theyre set to auto
    root@sled327857198-oob:~# mfg-tool fan-speed | tail -10
        },
        "FANBOARD1_FAN7": {
            "pwm": {
                "current": 100.0,
                "target": 100.0
            },
            "tach_inlet_speed_rpm": 15603.0,
            "tach_outlet_speed_rpm": 14456.0
        }
    }

Run the scrips without the control file and confirm that the fans are set to auto
    root@sled327857198-oob:~# bash -x /tmp/fan
    + FAN_MANUAL_MODE_ENV_FILE=/home/root/.fan_manual_mode.env
    + '[' '!' -f /home/root/.fan_manual_mode.env ']'
    + /usr/bin/mfg-tool fan-mode -a
    <7> Finding Control.Mode objects.
    <7> Iterating over entries.
    {}
    + exit 0

Touch the control file and run again. This sets the fans to 80%
{P1729834790}

Confirm afterwards
{P1729840626}

Reviewed By: patelhiral22

Differential Revision: D69425361

fbshipit-source-id: a33af3506072bb9a4f4163f828543d0474d9fc27