Don't blindly swallow exceptions
Summary:
We are seeing issues with the `swampd` service in which it is exhausting file descriptors. This is being swallowed by an exception handler that blindly ignores all exceptions and retries later.

In the case of FDs, we'll never get out of this stuck state.
{P1732200870}

While we investigate, we want the service to allow this exception to rise up and trigger the service to terminate, freeing all of the file descriptors.

Systemd will restart the service after 5s and we can watch the scuba logs for which exceptions need to be handled correctly.

Test Plan:
ran `devtool modify phosphor-pid-control` and then within the workspace,
ran `git log --pretty=oneline` and

  git log --pretty=oneline
  c3faf0a56958db50014c398fadafafbe7f15edc7 (HEAD -> devtool, tag: devtool-patched) Don't blindly swallow exceptions
  fc97b5cdd97cc73c8f98ec3cf194f368dbbe3965 (grafted, tag: devtool-base, origin/master, master) OWNERS: adjust Ed's email

we see the patch applied

This can't be tested in qemu, so I installed the binary to a sled and verified that it starts up normally.
{P1732203682}

Reviewed By: amithash

Differential Revision: D69608736

fbshipit-source-id: e10a5aed8415b1fd514a9ca41f5f5002eb8759df