Make sure we don't start sshd-remediate too early
Summary:
It looks like the first call to sshd-remediate always errors out and this gives
an error in our QEMU robot testing.

Secondly, there may be a bug without the delay in the first call that
errors out: when we error out the first time, we write a file
to /run. When we run this early in the startup process, mounts are being
made in that directory so there could be races in that inode.

Test Plan:
Made sure the QEMU error for journal FAIL logs from sshd-remediate are no
longer there. And ensure that sshd-remediate still runs.

Reviewed By: waffle2k

Differential Revision: D69793667

fbshipit-source-id: 45d71a76ceb19a20a26a45d52c7c9b0db733d0f0