dbus-sensors: fixed multiple power state issue
Summary:
fixed multiple power state issue:
If the entity does not set powerState for each channel, label_powerState will be contaminated by the state of the previous channel. Therefore, if label_powerState is not found,the powerState should be updated.

fixed not activate for multiple power state:
Due to the deactivation function being sensor-specific while the check findI2CDev != devices.end() is chip-specific, some sensors may fail to activate during power state changes.

For example, consider the following EM configuration:

{
  "in0_Name": "MB_P12V_MEM0_VOLT_V",
  "in0_PowerState": "Always",
  "in1_Name": "MB_P5V_STBY_VOLT_V",
  "in2_Name": "MB_PVDD33_S5_VOLT_V",
  "in3_Name": "MB_P1V8_STBY_VOLT_V",
  "in4_Name": "MB_P12V_MEM1_VOLT_V",
  "in5_Name": "MB_P1V8_CPU_RT_VOLT_V",
  "in5_PowerState": "On"
}

If in0_PowerState is set to Always, and we perform a DC off followed by a DC on, in5 might fail to activate. This happens because the condition "activateOnly && !findI2CDev->second.second" evaluates to true, causing the function to return early.

To address this issue, an additional condition has been added to bypass the check when multiple power states are detected, ensuring that all sensors can be properly activated

# Description

Please include a summary of the change and which issue is fixed.

# Motivation

Please include an explanation of why you these changes are necessary

X-link: https://github.com/facebookexternal/openbmc.quanta/pull/4573

Test Plan: Please describe the tests that you ran to verify your changes. Provide instructions so we can reproduce.

Reviewed By: amithash

Differential Revision: D70073466

fbshipit-source-id: dea6232f64be06c7c1dc71270e9eaf982bba6930