yosemite4: Add mem dump on warning
Summary:
As title, dump memory profile to /home/root/mem-out.txt
when we hit warning.

Keep this downstream till we figure out what info is generally
useful and we might want. Also the right thing to do is to have
this potentially trigger phosphor-debug-collector instead.
While we figure that out keeping this downstream.

Test Plan:
Load it on Qemu and check health configuration are updated.
run mem-dump service and ensure it creates a file in
/home/root/mem-out.txt

Reviewed By: waffle2k

Differential Revision: D70656013

fbshipit-source-id: bfd76a2f0b8af38528f51bdddc1bda2d7b3e7970