Elbert: Fix LED controld
Summary:
When sysv was converted to systemd, the led-controld daemon stopped running. As a result, LEDs on the supervisor were no longer being managed.

The fix is to implement a systemd service.

In addition, the CIT test test_process_running is enhanced to catch this problem in the future.

X-link: https://github.com/facebookexternal/openbmc.arista/pull/404

Test Plan:
1\. Ran CIT tests and ensured the test_process_running.py test passed.
2\. Verified the LEDs are now being lit on the supervisor of an Elbert.

Reviewed By: mikechoifb

Differential Revision: D71341317

Pulled By: joancaneus

fbshipit-source-id: d34f5a93afb402224c444b60e3c34de9c908782d