fix SIGSEGV in redfish-binding base class
Summary: Using a std::unordered_map to record the address of class member would cause SIGSEGV when the instance are copied, which will make all of the class member address change while the content of the map will just be copied over, which still have the old address. This diff fix this.

Test Plan:
build
```
bitbake catalina-image
```

Reviewed By: williamspatrick

Differential Revision: D70831004

fbshipit-source-id: b56308679c39365dadd8d72205896bbe86e64b70