Temp test to isolate memleak to SEL Logger handling
Summary:
As title, disable IPMI from handling added/removed
changes from phosphor-logging (We technically dont
use it anyway). But it gives us a way to know if
the issue is related to the components here.

Test Plan: Soaked it on a few systems overnight. Seems to have mem leak fixed. More testing pending.

Reviewed By: waffle2k

Differential Revision: D71369799

fbshipit-source-id: 11b7c194c9a13684ec00917ffc0ec3a4eb717829