redfish-client: eliminate generation of common.hpp
Summary:
We previously blindly copied `common.hpp` into the build tree
but none of the meson setup was there to add it as a dependency.  This
lead to confusion when editing the source file and not seeing those
changes surface in the build.

There isn't a strong reason to consider this a generated artifact.  We
can make a few minor changes and leave it as a static resource in the
source tree.

Related, there were test cases for these `common.hpp` property classes
which were never enabled.  Move those test cases to the test directory
and enable in `test/meson.build`.

Test Plan: Code compiles and unit tests pass.

Reviewed By: wangx6f

Differential Revision: D71525136

fbshipit-source-id: 73fda8acce75ebf6838ddafd32e3148e9ee76ddf