redfish-client: generate forEachProperty method
Summary:
In order for the `toJson` method to work, somehow we have to
generate the set of properties include in each generated class.  This
was previously missed.  Implement an override of the `forEachProperty`
method so that `toJson` is populated correctly with each property.

Test Plan: Functionality leveraged in subsequent commit.

Reviewed By: wangx6f

Differential Revision: D71525139

fbshipit-source-id: 113349ce0053ac8f4be821626481abef77930900