redfish-client: sensor: fix invalid stack reference lifetime
Summary:
I observed that the `dbus_server_tests` would fail
approximately 50% of the time.  This was traced down to a stack
reference being passed to a co-routine, which executed after the
stack was freed.

Co-routine functions generally should not take a reference, unless
that reference is known to exist longer than the co-routine execution
(such as the case of a reference to the program-level configuration
object in main).  Consider a co-routine such as:

```
    auto foo(Struct&& bar) -> task<void>;
```

It may seem normal to call this with `foo(std::move(value))`, but
calling `foo` creates a co-routine context and does not actually
execute anything in the co-routine.  This means that an r-value
reference is passed along to the co-routine.  If the co-routine is
executed after `value` is destructed, the co-routine has a reference
to a destructed object (likely on the stack).  For this reason,
co-routines almost always take values and not references.

Revamp the `daemon.cpp` to avoid passing `Sensor` instances by
reference.

Test Plan: Ran test cases 100% without fail.

Reviewed By: wangx6f

Differential Revision: D71642027

fbshipit-source-id: e9515c18f884c2d75f2dd9750d82ee67d7101137