change reference from "redfishlib" to "redfish-binding"
Summary:
In order to avoid confusion and naming consistency, this diff change the reference from "redfishlib" to "redfish-binding", mainly the namespace and the include header.

The reason why not call it "redfishlib" is mainly that the generated code is a collection of c++ class that modeling the redfish schema, but it does not handle any other protocol level logic such as redfish query.

Test Plan:
ptest
```
./tests2/ptest-runner redfish-client
```
build
```
source openbmc-init-build-env meta-facebook/meta-catalina build-catalina && bitbake catalina-image
```

Reviewed By: williamspatrick

Differential Revision:
D72149917

Privacy Context Container: L1188897

fbshipit-source-id: c01ebed5d25f5b7da953cf2dd2ed5612eb7ff91a