Add some retry and timeout logic
Summary:
This way if RTP wants to have retries, they can create a dotfile under `/home/root/.mfgtool.json`. This file is json file with key value pairs. Currently it support `retries` and `timeout`

If the command fails to execute, it will try again and again for the number defined in `retries`. It will exit once the rc of the child process exits cleanly, or the number of retries have been exhausted.

There's also a timeout feature. This simply creates an `alarm` signal and the child process is killed. The retries is honoured.

Test Plan:
Ran this on an actual BMC, populating the json file with timeouts and retry values. Watched it timeout appropriately as well as retried when it failed
{P1778970450}

Testing return codes
```
root@sled327803164-oob:~# /tmp/mfg-tool hello ; echo $?
<7> Hello World!
[
    "Hello World!"
]
0
root@sled327803164-oob:~# /tmp/mfg-tool helloXXX ; echo $?
A subcommand is required
Run with --help for more information.
106
```

Reviewed By: williamspatrick

Differential Revision: D72364358

fbshipit-source-id: 05c552c2ad7f3d3658ba3f9e9c702cc132cf7fa0