GT/GTI: gpio: rename "GPU_WP_HW_CTRL_N" gpio
Summary:
Correct the gpio name because it is a high-active GPIO. This gpio is used to in conjunction for the write protection of all EEPROM device on GPU board. The default value is low (Disabled).

X-link: https://github.com/facebookexternal/openbmc.quanta/pull/4631

Test Plan: Build and test pass on GTI system.

Reviewed By: Prabha-Veerubhotla

Differential Revision: D73644855

fbshipit-source-id: 57aae5e05d4d92e4c003023a3791fb6d4f77f820