Meru: Auto-fix DMI_BOARD_NAME
Summary:
This change automatically fixes the DMI_BOARD_NAME aboot_conf parameter based on what's programmed in the switchcard EEPROM.

One common scenario where DMI_BOARD_NAME may be mis-programmed is the scenario where the supervisor card is moved from one product to another. For example, when it's moved from Viper(Meru800bia) to ViperB(Meru800biab) or vice-versa.

On a cold boot (when both the BMC and CPU are power cycled), the aboot_conf section is queried and checked against the Product Name in the SMB eeprom. If the values are different and if the value in the SMB eeprom is valid, then the aboot_conf section is re-programmed.

This check is done prior to the CPU being powered on so as a running CPU would never be affected.

Since the entire aboot_conf section has to be programmed, the console speed is queried. If it's 57600, then the bmc configuration type is used. Otherwise, the cpu configuration type is used. This is to avoid changing the console speed.

X-link: https://github.com/facebookexternal/openbmc.arista/pull/407

Test Plan:
Implemented new dmidecode CIT test.

Verified on a Viper that the DMI_BOARD_NAME is fixed after changing the parameter and then doing a cold boot.

Reviewed By: mikechoifb

Differential Revision: D73673492

Pulled By: joancaneus

fbshipit-source-id: bb53bd6f7ea07782712a32e6680dcf9d39936d84