greatlakes-lf: delete machine
Summary:
The upstream tree for greatlakes was deleted because it was
a dead experiment.  We need to delete our downstream code as well.
Right now, Yocto syncs are unsuccessful because the upstream meta-layer
no longer exists but the machine is referenced by the
openbmc-init-build-env.

Since the platform is already marked deprecated, there should be no
existing build use cases of this code.

Test Plan: Confirm build passes in CIT.

Reviewed By: jamesatha

Differential Revision: D75041317

fbshipit-source-id: b9634e79af16ff3b862a1c71d9103ee9b9e5ea9b