Allow duplicate addresses
Summary:
This removes the hardcoded assumption that a device with a specific
address will only appear on one bus. With this change, we introduce a
port to distinguish different buses. Because we want this to work with
current systems (where this is not specified) and new, we have made this
an optional configuration and field for the various APIs

Test Plan:
All tests in `./unit` succeed. We will run this on an RMC to
make sure it works as intended (still waiting on hardware for that)

Installed this on a ventura and ran `rackmoncli list`:
```
baudrate     | crcErrors    | devAddress   | deviceType   | miscErrors   | mode         | timeouts     | uniqueDevAddress |
-----------------------------------------------------------------------------------------------------------------------
19200        | 0            | 12           | ORV3_RPU     | 0            | ACTIVE       | 0            | 268          |
19200        | 0            | 12           | ORV3_RPU     | 0            | ACTIVE       | 0            | 524          |

```

Reviewed By: amithash

Differential Revision:
D72617675

Privacy Context Container: L1188897

fbshipit-source-id: 839f8c5817f8799c414b15980042f6dba36cfea4