yosemite4: spec: add mfg-tool.md to COMMON.yaml
Summary:
# Description:
This commit adds a new description file `mfg-tool.md` to the miscellaneous directory, and links it in the `COMMON.yaml` file.

# Motivation:
Adding the `mfg-tool.md` file provides information about the usage and functionality of the manufacturing tool, which is essential for users to operate the tool properly.

X-link: https://github.com/facebookexternal/openbmc.wiwynn/pull/3316

Reviewed By: jamesatha, waffle2k

Differential Revision: D74865278

fbshipit-source-id: 042fbe922039ae061b9d3681a099cf220cd4d690